refactor(settings): extract sections into reusable components
Decomposes the ~840-line Settings page into per-section components under
components/settings/, reconciling with the drifted standalone Section
components and removing the duplicated inline primitives.

- Extract Profile, Notifications, Security, Wallet, Family and Preferences
  sections plus shared primitives (SectionCard, SectionHeader, FieldRow,
  TextInput, Toggle, SaveButton) and the InsuranceReminderPreview widget.
- Page shell keeps SECTIONS as the single source of nav order plus the
  IntersectionObserver scroll-spy; section ids stay in sync.
- Delete unused, drifted duplicates: components/SecuritySection.tsx,
  AccountSection.tsx, SettingsSection.tsx, PreferencesRow.tsx.
- Preserve useDensity() integration, i18n, a11y and styling exactly
  (pure decomposition, no behavior change).
- Add focused component tests for each section + primitives + the
  insurance preview; renderWithProviders now also supplies ToastProvider.
- Document the structure in docs/settings-page-structure.md.
