+++ tests/react/renderWithProviders.tsx
@@ -1,6 +1,7 @@
 import React from "react";
 import { render } from "@testing-library/react";
 import { DensityProvider, type Density } from "@/lib/context/DensityContext";
+import { ToastProvider } from "@/lib/context/ToastContext";
 
 export function renderWithProviders(
   ui: React.ReactElement,
@@ -14,5 +15,9 @@ export function renderWithProviders(
   // jsdom localStorage is available.
   window.localStorage.setItem("display-density", density);
 
-  return render(<DensityProvider>{ui}</DensityProvider>);
+  return render(
+    <DensityProvider>
+      <ToastProvider>{ui}</ToastProvider>
+    </DensityProvider>,
+  );
 }
